chatbot_eval_git/requirements.txt
torch
transformers
huggingface-hub
numpy
Flask
Flask-CORS
mysql-connector-python
requests
pandas 
openpyxl 
sentence-transformers 
accelerate

